Register the outputs of non-resolved derivations
When building a (non fully-input-addressed) derivation, we first resolve it to
a `BasicDerivation` that depends directly on the output paths of its
dependencies (rather than on the derivation outputs) − this is what enables
early cutoff.

Currently we register the realisation of the resolved derivation, so that
subsequent calls to `nix-build` can directly reuse this without rebuilding the
derivation.  However, we don't register the unresolved derivation, meaning that
we have to re-do the resolution step each time. This in turn means that we must
either keep all the build inputs or keep the knowledge of their output path
(without the content of the path itself). The former is very costly in
disk-space (it amounts to making every build-input a runtime dependency), and
the latter comes with a big set of challenges in presence of non-determinism
that makes this a too heavy-handed solution for the time being (though we might
switch to it eventually if we can find a way to solve its issues as it's a
rather tempting design).

To avoid these issues, we choose a third way which is to also register the
realisation of the unresolved derivation − so we can completely sidestep the
resolving phase, which makes these behave as classical input-addressed
derivations.